refactor(GroupBy): move to ui/model
