+++ app/src/main/java/com/augieafr/todo_app/ui/model/GroupBy.kt
@@ -1,0 +1,5 @@
+package com.augieafr.todo_app.ui.model
+
+enum class GroupBy {
+    DEADLINE, IS_DONE
+}
